Move BuildConfig to data layer
