+++ app/build.gradle
@@ -3,9 +3,6 @@ plugins {
     id 'org.jetbrains.kotlin.android'
 }
 
-Properties properties = new Properties()
-properties.load(project.rootProject.file('local.properties').newDataInputStream())
-
 android {
     compileSdk 32
 
@@ -17,8 +14,6 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        buildConfigField "String", "Weather_Api_Address", properties['weather_api_address']
-        buildConfigField "String", "Weather_Api_Key", properties['weather_api_key']
     }
 
     buildTypes {
