+++ data/src/main/java/com/example/bcsd_weather/data/api/UltraShortTermForecastInterface.kt
@@ -1,6 +1,6 @@
 package com.example.bcsd_weather.data.api
 
-import com.example.bcsd_weather.BuildConfig
+import com.example.bcsd_weather.data.BuildConfig
 import com.example.bcsd_weather.data.model.UltraShortTermForeCastRemote
 import retrofit2.http.GET
 import retrofit2.http.Query
